requirements.txt
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
supabase==1.2.0
